pyproject.toml
[tool.poetry]
name = "nwav"
version = "0.1.0"
description = ""
authors = ["ricepotato <ricepotato40@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
